// src/usecase/customer/create/create.customer.usecase.ts
import Customer from "../../../domain/customer/entity/customer";
import CustomerRepositoryInterface from "../../../domain/customer/repository/customer-repository.interface";
import Address from "../../../domain/customer/value-object/address";
import { InputCreateCustomerDTO, OutputCreateCustomerDTO } from "./create.customer.dto";
import { v4 as uuid } from "uuid";


export default class CreateCustomerUseCase {

    private customerRepository: CustomerRepositoryInterface;

    constructor(customerRepository: CustomerRepositoryInterface) {
        this.customerRepository = customerRepository;        
    }

    async execute(input: InputCreateCustomerDTO): Promise<OutputCreateCustomerDTO> {
        const customerId = uuid();
        
        const customer = new Customer(customerId, input.name);
        const address = new Address(input.address.street, input.address.number, input.address.zip, input.address.city);
        customer.changeAddress(address);

        await this.customerRepository.create(customer);

        const retorno = {
            id: customer.id,
            name: customer.name,
            address: {
                street: customer.address.street,
                city: customer.address.city,
                number: customer.address.number,
                zip: customer.address.zip
            }
        }
        return retorno;
    }
}